Restructure of code, addition of more cogs.
Restructured cogs to make it look nicer and easier to read. Added load, unload, reload function so that we don't have to rerun the bot every time. Added game cog, but is only a prototype for what's to come. Got rid of valorant cog since I might do TFT or just redo it.